Added test for Ai, PlayerConroller, and Player classes. Fixed errors that showed up through testing.

+++ src/Player.java
@@ -2,7 +2,7 @@ public class Player extends PlayerController {
 
     private boolean continuePlaying;
 
-    public Player(boolean isAi, byte lives) {
+    public Player(boolean isAi, int lives) {
         super(isAi, lives);
     }
